# openapi.json
{
  "components": {
    "schemas": {
...
      "CareItemExpirationResponse": {
        "properties": {
          "care_item_id": {
            "format": "uuid",
            "title": "Care Item Id",
            "type": "string"
          },
          "expiration_date": {
            "format": "date",
            "title": "Expiration Date",
            "type": "string"
          }
        },
        "required": [
          "care_item_id",
          "expiration_date"
        ],
        "title": "CareItemExpirationResponse",
        "type": "object"
      },
      "CareItemExpirationUpdateRequest": {
        "properties": {
          "expiration_date": {
            "format": "date",
            "title": "Expiration Date",
            "type": "string"
          }
        },
        "required": [
          "expiration_date"
        ],
        "title": "CareItemExpirationUpdateRequest",
        "type": "object"
      },
...
      "CareItemListResponse": {
        "examples": [
          {
            "has_next": false,
            "items": [
              {
                "brand": "Life Extension",
                "created_at": "2026-08-13T09:00:00Z",
                "days_until_depletion": 28,
                "dose_per_intake": "1",
                "expected_depletion_date": "2026-09-10",
                "id": "31000000-0000-4000-8000-000000000001",
                "image_url": "/static/products/life-extension-two-per-day.svg",
                "intake_start_date": "2026-08-12",
                "intakes_per_day": 2,
                "name": "라이프익스텐션 투퍼데이",
                "product_id": "22000000-0000-4000-8000-000000000001",
                "product_type": "SUPPLEMENT",
                "purchase_date": "2026-08-10",
                "quantity_unit": "CAPSULE",
                "total_quantity": "60"
              }
            ],
            "page": 1,
            "page_size": 20,
            "total": 1
          }
        ],
        "properties": {
          "has_next": {
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CareItemListItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "has_next"
        ],
        "title": "CareItemListResponse",
        "type": "object"
      },
...
      "DailyIntakeResponse": {
        "examples": [
          {
            "nutrients": [
              {
                "daily_amount": "470",
                "nutrient_code": "VITAMIN_C",
                "nutrient_id": "23000000-0000-4000-8000-000000000001",
                "nutrient_name": "비타민 C",
                "unit": "MG"
              }
            ]
          }
        ],
        "properties": {
          "nutrients": {
            "items": {
              "$ref": "#/components/schemas/DailyIntakeNutrientResponse"
            },
            "title": "Nutrients",
            "type": "array"
          }
        },
        "required": [
          "nutrients"
        ],
        "title": "DailyIntakeResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AccessCookieAuth": {
        "description": "HttpOnly access JWT 쿠키",
        "in": "cookie",
        "name": "yeongyangkkuk_access_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "영양꾹 1차 로컬 MVP 백엔드 API",
    "title": "영양꾹 백엔드",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/care/daily-intake": {
      "get": {
        "operationId": "care_get_daily_intake",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyIntakeResponse"
                }
              }
            },
            "description": "활성 영양제 복용 계획의 성분별 일일 예정 섭취량",
            "headers": {
              "Cache-Control": {
                "description": "개인 건강 관련 응답 캐시 금지",
                "schema": {
                  "example": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTH_REQUIRED",
                    "fields": [],
                    "message": "인증이 필요합니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "access 인증 실패"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SERVICE_UNAVAILABLE",
                    "fields": [],
                    "message": "서비스가 아직 준비되지 않았습니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PostgreSQL 조회 또는 단위 변환 실패"
          }
        },
        "security": [
          {
            "AccessCookieAuth": []
          }
        ],
        "summary": "내 일일 예정 섭취량 조회",
        "tags": [
          "마이케어"
        ]
      }
    },
    "/api/v1/care/items": {
      "get": {
        "operationId": "care_list_items",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CareItemListResponse"
                }
              }
            },
            "description": "현재 사용자의 활성 복용 제품 페이지",
            "headers": {
              "Cache-Control": {
                "description": "개인 건강 관련 응답 캐시 금지",
                "schema": {
                  "example": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTH_REQUIRED",
                    "fields": [],
                    "message": "인증이 필요합니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "access 인증 실패"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SERVICE_UNAVAILABLE",
                    "fields": [],
                    "message": "서비스가 아직 준비되지 않았습니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PostgreSQL 조회 실패"
          }
        },
        "security": [
          {
            "AccessCookieAuth": []
          }
        ],
        "summary": "내 활성 복용 제품 목록 조회",
        "tags": [
          "마이케어"
        ]
      },
      "post": {
        "operationId": "care_register_item",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CareItemCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CareItemResponse"
                }
              }
            },
            "description": "복용 제품 등록 성공",
            "headers": {
              "Cache-Control": {
                "description": "개인 건강 관련 응답 캐시 금지",
                "schema": {
                  "example": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTH_REQUIRED",
                    "fields": [],
                    "message": "인증이 필요합니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "access 인증 실패"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "PRODUCT_NOT_FOUND",
                    "fields": [],
                    "message": "제품을 찾을 수 없습니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "카탈로그 제품 없음"
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "VALIDATION_FAILED",
                    "fields": [],
                    "message": "요청 값을 확인해주세요.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "등록 값 검증 실패"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SERVICE_UNAVAILABLE",
                    "fields": [],
                    "message": "서비스가 아직 준비되지 않았습니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PostgreSQL 처리 실패"
          }
        },
        "security": [
          {
            "AccessCookieAuth": []
          }
        ],
        "summary": "복용 제품 등록",
        "tags": [
          "마이케어"
        ]
      }
    },
    "/api/v1/care/items/{care_item_id}": {
      "delete": {
        "operationId": "care_delete_item",
        "parameters": [
          {
            "in": "path",
            "name": "care_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Care Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "복용 항목 소프트 삭제 성공",
            "headers": {
              "Cache-Control": {
                "description": "개인 건강 관련 응답 캐시 금지",
                "schema": {
                  "example": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTH_REQUIRED",
                    "fields": [],
                    "message": "인증이 필요합니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "access 인증 실패"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CARE_ITEM_NOT_FOUND",
                    "fields": [],
                    "message": "복용 항목을 찾을 수 없습니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "현재 사용자의 활성 복용 항목 없음"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SERVICE_UNAVAILABLE",
                    "fields": [],
                    "message": "서비스가 아직 준비되지 않았습니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PostgreSQL 처리 실패"
          }
        },
        "security": [
          {
            "AccessCookieAuth": []
          }
        ],
        "summary": "내 복용 제품 삭제",
        "tags": [
          "마이케어"
        ]
      }
    },
    "/api/v1/care/items/{care_item_id}/expiration": {
      "put": {
        "operationId": "care_update_item_expiration",
        "parameters": [
          {
            "in": "path",
            "name": "care_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Care Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CareItemExpirationUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CareItemExpirationResponse"
                }
              }
            },
            "description": "현재 사용자의 활성 복용 항목 유통기한 갱신",
            "headers": {
              "Cache-Control": {
                "description": "개인 건강 관련 응답 캐시 금지",
                "schema": {
                  "example": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTH_REQUIRED",
                    "fields": [],
                    "message": "인증이 필요합니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "access 인증 실패"
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "CARE_ITEM_NOT_FOUND",
                    "fields": [],
                    "message": "복용 항목을 찾을 수 없습니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "현재 사용자의 활성 복용 항목 없음"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SERVICE_UNAVAILABLE",
                    "fields": [],
                    "message": "서비스가 아직 준비되지 않았습니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "PostgreSQL 처리 실패"
          }
        },
        "security": [
          {
            "AccessCookieAuth": []
          }
        ],
        "summary": "내 복용 제품 유통기한 추가·교정",
        "tags": [
          "마이케어"
        ]
      }
    },
    "/api/v1/care/nutrient-status": {
      "get": {
        "operationId": "care_get_nutrient_status",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NutrientStatusResponse"
                }
              }
            },
            "description": "등록된 보충제 복용 계획이 총 식이 기준량에서 차지하는 비율. 실제 음식 섭취량이나 임상적 결핍·과잉 판정이 아닙니다.",
            "headers": {
              "Cache-Control": {
                "description": "개인 건강 관련 응답 캐시 금지",
                "schema": {
                  "example": "no-store",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "AUTH_REQUIRED",
                    "fields": [],
                    "message": "인증이 필요합니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "access 인증 실패"
          },
          "503": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SERVICE_UNAVAILABLE",
                    "fields": [],
                    "message": "서비스가 아직 준비되지 않았습니다.",
                    "request_id": "opaque-request-id"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "기준 버전 또는 PostgreSQL 조회·단위 검증 실패"
          }
        },
        "security": [
          {
            "AccessCookieAuth": []
          }
        ],
        "summary": "내 영양성분 현황 조회",
        "tags": [
          "마이케어"
        ]
      }
    },
...
  }
}